ccfri-4682 - move the application template check from routes to components
