+++ frontend/src/components/applicationTemplates/v1/group/CCOF/FacilityInformation.vue
