+++ frontend/src/components/applicationTemplates/v1/group/summary/OrganizationSummary.vue
